Add, Refactor : 회원 탈퇴 기능 추가, 일부 메서드의 클래스 위치 변경
