+++ src/main/java/techcourse/myblog/domain/UserRepository.java
@@ -8,4 +8,6 @@ public interface UserRepository extends CrudRepository<User, Long> {
     List<User> findUsersByEmail(String email);
 
     User findUserByEmail(String email);
+
+    User findUserById(long id);
 }
