+++ src/main/resources/templates/fragments/logged-in.html
@@ -11,7 +11,7 @@
         </a>
         <ul class="dropdown-menu">
             <li class="text-dark">
-                <a href="mypage.html">
+                <a href="/mypage">
                     <i class="ti-user pdd-right-10"></i>
                     <span class="text-dark">마이 페이지</span>
                 </a>
