+++ src/main/resources/templates/mypage.html
@@ -99,7 +99,7 @@
                                 <p class="mrg-top-10 text-dark"> <b>Name</b></p>
                             </div>
                             <div class="col-md-6">
-                                <p class="mrg-top-10">Brown</p>
+                                <p class="mrg-top-10" th:text="${session.user.name}"></p>
                             </div>
                         </div>
                         <hr>
@@ -108,7 +108,7 @@
                                 <p class="mrg-top-10 text-dark"> <b>Email</b></p>
                             </div>
                             <div class="col-md-6">
-                                <p class="mrg-top-10">gracefulBrown@woowahan.com</p>
+                                <p class="mrg-top-10" th:text="${session.user.email}"></p>
                             </div>
                         </div>
                         <hr>
@@ -158,12 +158,28 @@
                                 <hr>
                             </div>
                         </div>
+                        <form th:action="'/user/delete/' + ${session.user.id}" id="deleteForm" method="post">
+                            <input th:if="${session.user}" type="hidden" name="_method" value="delete">
+                            <button type="button" id="deleteButton" class="btn btn-light" onclick="deleteUser()">Leave</button>
+                        </form>
                     </div>
                 </div>
             </div>
         </div>
     </div>
 </div>
+
+<script >
+    var deleteForm = document.getElementById("deleteForm");
+    var result;
+
+    function deleteUser() {
+        result = confirm("정말로 탈퇴하시겠습니까");
+        if (result === true) {
+            deleteForm.submit();
+        }
+    }
+</script>
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.bundle.min.js" integrity="sha384-pjaaA8dDz/5BgdFUPX6M/9SUZv4d12SUPF0axWc+VRZkx5xU3daN+lYb49+Ax+Tl" crossorigin="anonymous"></script>
